Move pprof and webserver into util
