+++ cmd/tsdb-gw/main.go
@@ -129,6 +129,7 @@ func main() {
 	initRoutes(api, *enforceRoles)
 
 	ms := util.NewMetricsServer(*metricsAddr)
+	util.StartDebugServer()
 
 	log.Infof("Starting %v ...", app)
 	done := make(chan struct{})
